refactor(#6a): split bot_arb.py into execute + scan mixins (1058->324)
Pure structural move (Phase-6 style). The per-arb-type executors (negrisk /
internal / cross-exchange) go to bot_arb_execute.py as ArbTradeExecutionMixin;
the scan/drive loop + depth research go to bot_arb_scan.py as ArbScanLoopMixin.
ArbExecutionMixin now inherits both and keeps only the conditional-arb +
rebalance helpers — so AuramaurBot's mixed-in method set is byte-identical and
the shared-self contract is preserved. bot_arb.py 1058->324; both new modules
added to the no-unbound-names AST guard (the console/log free-name trap).

Full suite 896 passed; ruff clean.
